[edit]: 와글리 게시물 만들기, 게시물 수정 분리 작업 완료
