+++ lib/controller/post/image_controller.dart
@@ -31,7 +31,6 @@ class ImageController extends GetxController {
         }
       }
     }
-
     showImages.add(images);
   }
