+++ lib/main.dart
@@ -22,13 +22,12 @@ import 'package:waggly/screens/chat/chat_room_screen.dart';
 import 'package:waggly/screens/chat/chat_search.dart';
 import 'package:waggly/screens/groupChat/group_chat_create.dart';
 import 'package:waggly/screens/post/edit.dart';
-import 'package:waggly/screens/post/writeTest.dart';
+import 'package:waggly/screens/post/write.dart';
 import 'package:waggly/screens/user/find_password.dart';
 import 'package:waggly/screens/index.dart';
 import 'package:waggly/screens/user/my_page.dart';
 import 'package:waggly/screens/post/post_home.dart';
 import 'package:waggly/screens/user/sign_in.dart';
-import 'package:waggly/screens/post/write.dart';
 import 'package:waggly/components/myPage/profileImg/profile_img.dart';
 import 'package:waggly/components/myPage/active/index.dart';
 import 'screens/post/post_detail_screen.dart';
@@ -110,7 +109,7 @@ class HeroApp extends StatelessWidget {
             GetPage(name: "/signInPage", page: () => SignInScreen(), transition: Transition.rightToLeft),
             GetPage(
               name: "/writePage",
-              page: () => WritePage2("write"),
+              page: () => WritePage("write"),
               transition: Transition.rightToLeft,
               binding: BindingsBuilder<PostController>(() {
                 Get.reload<PostController>();
