+++ lib/screens/post/write.dart
@@ -62,17 +62,7 @@ class WritePage extends StatelessWidget {
     var _page = Status.edit;
     var _os = Platform.operatingSystem;
     var _postName = "게시글 작성";
-    if (type == "edit") {
-      _postName = "게시글 수정";
-      final PostDetailController postDetailController = Get.put(PostEditController());
-      _imageController.getImagesUrl(postDetailController.postDetail.value.postImages);
-      _title = TextEditingController(text: postDetailController.postDetail.value.postTitle);
-      _content = TextEditingController(text: postDetailController.postDetail.value.postDesc);
-    } else {
-      _imageController.getImagesUrl([]);
-      _title = TextEditingController();
-      _content = TextEditingController();
-    }
+    _imageController.getImagesUrl([]);
 
     return Scaffold(
       resizeToAvoidBottomInset: true,
@@ -142,10 +132,7 @@ class WritePage extends StatelessWidget {
                 child: Obx(
                   () => PhotoWidget(
                       imageController: _imageController,
-                      imageUrlLength: _imageController.imagesUrl.length,
-                      imagesLength: _imageController.images!.length,
-                      length: _imageController.showImages.length,
-                      type: "write"),
+                      imagesLength: _imageController.images!.length),
                 ),
               ), // 선택된 이미지 표시 영역
               Padding(
@@ -195,52 +182,16 @@ class WritePage extends StatelessWidget {
                       width: double.infinity,
                       height: _bottomButtonHeight.h,
                       decoration: BoxDecoration(
-                        color: _postController.isButtonActivate.value == true ? Palette.main : Palette.paper,
+                        color: _postController.isButtonActivate.value == true
+                            ? Palette.main
+                            : Palette.paper,
                         borderRadius: BorderRadius.circular(40),
                       ),
                       child: TextButton(
-                        onPressed: type == "edit"
-                            ? () async {
-                                await editPost(_title.text, _content.text, "SOCIAL");
-                              }
-                            : _postController.isButtonActivate.value == true
-                                ? () async {
-                                    final result = await writePost();
-                                    print("나를 채킹하세요");
-                                    print(result);
-                                    if (result.code == 201) {
-                                      _postController.isButtonActivate.value = false;
-                                      CustomSnackBar.messageSnackbar(
-                                        context,
-                                        "게시글 작성이 완료되었습니다.",
-                                        EdgeInsets.only(bottom: 65.h, left: 20.w, right: 20.w),
-                                      );
-                                      await Future.delayed(const Duration(seconds: 2), () {
-                                        Get.offAllNamed("/");
-                                      });
-                                    } else {
-                                      CustomSnackBar.messageSnackbar(
-                                        context,
-                                        result.message,
-                                        EdgeInsets.only(bottom: 65.h, left: 20.w, right: 20.w),
-                                      );
-                                    }
-                                  }
-                                : () {
-                                    if (_title.text.isEmpty) {
-                                      CustomSnackBar.messageSnackbar(
-                                        context,
-                                        "제목을 입력해주세요.",
-                                        EdgeInsets.only(bottom: 65.h, left: 20.w, right: 20.w),
-                                      );
-                                    } else if (_content.text.isEmpty) {
-                                      CustomSnackBar.messageSnackbar(
-                                        context,
-                                        "내용을 입력해주세요.",
-                                        EdgeInsets.only(bottom: 65.h, left: 20.w, right: 20.w),
-                                      );
-                                    }
-                                  },
+                        onPressed: () {
+                          final result = writePost();
+                          print(result);
+                        },
                         child: Text(
                           "게시글 작성하기",
                           style: _postController.isButtonActivate.value == true
@@ -268,21 +219,25 @@ class WritePage extends StatelessWidget {
   }
 
   Future<void> editPost(title, description, college) async {
-    final PostDetailController postDetailController = Get.put(PostEditController());
+    final PostDetailController postDetailController =
+        Get.put(PostEditController());
 
     List<MultipartFile> file = imageToMultipartFile();
-    await postDetailController
-        .editBoard(PostEditRequestDto(title, description, college, file, _imageController.parseToStringList()));
+    await postDetailController.editBoard(PostEditRequestDto(title, description,
+        college, file, _imageController.parseToStringList()));
   }
 
   Future<WagglyResponseDto> writePost() async {
     List<MultipartFile> file = imageToMultipartFile();
     List<String> hashtags = extractHashTags(_hashtag.text);
+    final box = Hive.box<User>('user');
+    int? me = box.get('user')?.id;
+
     return await _postController.writeBoard(
       PostRequestDto(
         _title.text,
         _content.text,
-        "ENGINEERING",
+        "SOCIAL",
         _postController.checkBox.value,
         hashtags,
         file,
@@ -295,22 +250,15 @@ class PhotoWidget extends StatelessWidget {
   const PhotoWidget(
       {Key? key,
       required this.imageController,
-      required this.imagesLength,
-      required this.imageUrlLength,
-      required this.length,
-      required this.type})
+      required this.imagesLength})
       : super(key: key);
 
-  final int length;
   final int imagesLength;
-  final int imageUrlLength;
   final ImageController imageController;
-  final String type;
 
   @override
   Widget build(BuildContext context) {
-    return length > 0
-        ? SizedBox(
+    return SizedBox(
             height: _imageThumbnailAreaHeight.h,
             child: ListView.separated(
               separatorBuilder: (BuildContext context, int index) {
@@ -319,88 +267,49 @@ class PhotoWidget extends StatelessWidget {
                 );
               },
               scrollDirection: Axis.horizontal,
-              itemCount: imageUrlLength + imagesLength,
+              itemCount: imagesLength,
               itemBuilder: (ctx, index) {
-                if (index < imageUrlLength) {
-                  return Row(
-                    children: [
-                      // if (index == 0) SizedBox(width: 20.0.w),
-                      Container(
-                        margin: EdgeInsets.only(bottom: 10.0.w),
-                        height: 100.0.h,
-                        width: 100.0.w,
-                        child: Stack(
-                          fit: StackFit.expand,
-                          children: [
-                            ClipRRect(
-                              borderRadius: BorderRadius.circular(18.0),
-                              child: Image.network(
-                                imageController.imagesUrl[index],
-                                fit: BoxFit.fill,
+                return Row(
+                  children: [
+                    Container(
+                      margin: EdgeInsets.only(bottom: 10.0.w),
+                      height: 100.0.h,
+                      width: 100.0.w,
+                      child: Stack(
+                        fit: StackFit.expand,
+                        children: [
+                          ClipRRect(
+                            borderRadius: BorderRadius.circular(18.0),
+                            child: Image.file(
+                              File(imageController
+                                  .images![index].path),
+                              fit: BoxFit.fill,
+                            ),
+                          ),
+                          Positioned(
+                            top: 6.h,
+                            left: _imageThumbnailAreaHeight - 27.w,
+                            child: InkWell(
+                              onTap: () {
+                                print(index);
+                                imageController.images?.removeAt(index);
+                              },
+                              child: SvgPicture.asset(
+                                "assets/icons/cancel.svg",
+                                width: 14.0.w,
+                                height: 14.0.h,
                               ),
                             ),
-                            Positioned(
-                              top: 6.h,
-                              left: _imageThumbnailAreaHeight - 27.w,
-                              child: InkWell(
-                                onTap: () {
-                                  imageController.deleteImages.add(imageController.imagesUrl[index]);
-                                  imageController.imagesUrl.removeAt(index);
-                                },
-                                child: SvgPicture.asset(
-                                  "assets/icons/cancel.svg",
-                                  width: 14.0.w,
-                                  height: 14.0.h,
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
+                          ),
+                        ],
                       ),
-                    ],
-                  );
-                } else {
-                  return Row(
-                    children: [
-                      Container(
-                        margin: EdgeInsets.only(bottom: 10.0.w),
-                        height: 100.0.h,
-                        width: 100.0.w,
-                        child: Stack(
-                          fit: StackFit.expand,
-                          children: [
-                            ClipRRect(
-                              borderRadius: BorderRadius.circular(18.0),
-                              child: Image.file(
-                                File(imageController.images![index - imageUrlLength].path),
-                                fit: BoxFit.fill,
-                              ),
-                            ),
-                            Positioned(
-                              top: 6.h,
-                              left: _imageThumbnailAreaHeight - 27.w,
-                              child: InkWell(
-                                onTap: () {
-                                  imageController.showImages[1].removeAt(index - imageUrlLength);
-                                },
-                                child: SvgPicture.asset(
-                                  "assets/icons/cancel.svg",
-                                  width: 14.0.w,
-                                  height: 14.0.h,
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                      // if (index == imagesLength - 1) SizedBox(width: 20.0.w),
-                    ],
-                  );
-                }
+                    ),
+                    // if (index == imagesLength - 1) SizedBox(width: 20.0.w),
+                  ],
+                );
               },
             ),
-          )
-        : SizedBox(height: _imageThumbnailAreaHeight.h);
+    );
   }
 }
 
@@ -425,7 +334,7 @@ class TopAppBar extends StatelessWidget with PreferredSizeWidget {
             crossAxisAlignment: CrossAxisAlignment.center,
             children: [
               InkWell(
-                onTap: (){
+                onTap: () {
                   Navigator.pop(context);
                 },
                 child: Container(
@@ -446,7 +355,9 @@ class TopAppBar extends StatelessWidget with PreferredSizeWidget {
               SizedBox(
                 width: 8.w,
               ),
-              Container(padding: EdgeInsets.only(bottom: 3.h), child: Text( "게시글 작성", style: CommonText.BodyL))
+              Container(
+                  padding: EdgeInsets.only(bottom: 3.h),
+                  child: Text("게시글 작성", style: CommonText.BodyL))
             ],
           ),
         ),
